add functionality to models and add tests

+++ choose_your_own_adventure/model/adventure_model.php
@@ -1,45 +1,67 @@
 <?php
-    include("app_model.php");
+    include_once "app_model.php";
     
     class AdventureModel extends AppModel
     {
         // constructor
-        public function __construct($date_created, $title, $user_id)
+        public function __construct($title, $user_id, $date_created = NULL)
         {
             parent::__construct($date_created);
             
             $this->fields["title"] = array($title, "VARCHAR(512)");
-            $this->fields["user_id"] = array($user_id, "INT(11), FOREIGN KEY(user_id) REFERENCES user_models(id)");
+            $this->fields["user_id"] = array($user_id, "INT(11), FOREIGN KEY(user_id) REFERENCES user_models(id) ON DELETE CASCADE");
             parent::construct_if_not_exists();
         }
         
         public static function find($id)
         {
-            $res = parent::find($id, AdventureModel::get_table_name());
-            $am = new AdventureModel($res["date_created"], $res["title"], $res["user_id"]);
-            $am->id = $res["id"];
+            $row = parent::find($id, AdventureModel::get_table_name());
+            $am = new AdventureModel($row["title"], $row["user_id"], $row["date_created"]);
+            $am->id = $row["id"];
+            return $am;
+        }
+        
+        public static function first()
+        {
+            $row =  parent::first(AdventureModel::get_table_name());
+            $am = new AdventureModel($row["title"], $row["user_id"], $row["date_created"]);
+            $am->id = $row["id"];
+            return $am;
+        }
+        
+        public static function last()
+        {
+            $row =  parent::last(AdventureModel::get_table_name());
+            $am = new AdventureModel($row["title"], $row["user_id"], $row["date_created"]);
+            $am->id = $row["id"];
             return $am;
         }
     }
     
     //Test
-    function test()
+    function adventure_model_test($delete)
     {
-        $adv = new AdventureModel(date("Y-m-d H:i:s"), "You shouln't see me" . rand(0, 100000), 3);
+        include "user_model.php";
+        $adv = new AdventureModel("Sample Title_" . rand(0, 100000), UserModel::last()->id, date("Y-m-d H:i:s"));
         $adv->save();
         $adv->print_fields();
         
-        $adv->set("title", "This is a new title");
+        $adv->set("title", "This is a new title buddy!");
         $adv->save();
         $adv->print_fields();
-        $adv->delete();
-    
-         $am = AdventureModel::find(3);
-         $am->print_fields();
+        
+        if($delete)
+            $adv->delete();
+        
+        $am = AdventureModel::find(AdventureModel::first()->id);
+        $am->print_fields();
          
-         AdventureModel::find(999);
+        AdventureModel::find(999);
+    }    
+
+    if(isset($argv[1]))
+    {
+        $delete = isset($argv[2]) ? FALSE : TRUE;
+        adventure_model_test($delete);
     }
-    
-    if($argv[1])
-        test();
 ?>
